requirements.txt
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0
opencv-python>=4.6.0
pillow>=8.3.2
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
PyQt5>=5.15.0
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0
requests>=2.25.0
roboflow>=1.1.0
pyyaml>=6.0